Add an HTTP interface to the most basic stats.
Just to get a feel for how this might work. Right now, you can get (as
JSON) the number of active message batches in the buffer (i.e. number of
distinct summary emails that are being built up and haven't yet been
sent) and the total number of received messages associated with those.

This is useful for manual testing, but probably not much else at this
point.