Send one summary email per recipient.
Prior to this commit, each received message would be rolled into
summaries by a key, and the recipients of the summaries would be derived
from the union of the original recipients of all the summarized
messages. For example: a message goes to both A and B, and another
message with the same key goes to just A; one summary of both messages
is sent, addressed to both A and B.

Now, messages are bucketed by recipients, and a single incoming message
with multiple recipients gets added to one bucket per recipient. So if a
message goes to both A and B, and another message with the same key goes
to just A; one summary of both messages is sent, addressed to just A,
and one summary of only the first message is sent, addressed to just B.