Rework the MessageStore interface.
This is most of the work required to have receiving/storing messages
happen in a separate process from summarizing and sending. The old
interface didn't make it easy to support a disk-backed store that wasn't
just a hack on top of the original memory store.

One particularly important result of all this is that the "batching
function" (that determines how messages are broken up into summaries) is
now applied _after_ storage rather than before -- i.e., it will now be
applied by the summarizer/sender, rather than the receiver.

This slightly improves the way Maildir was being used for metadata
storage too -- which helps the disk-backed storage implementation better
conform to the new interface.

Also, note that the memory store is still around, mostly used in tests.
It might get fully ripped out at some later point.

There are still a bunch of issues this doesn't address, but it should
make the commits that address them much more coherent.