Add a first cut at address rewriting.
This lets us have failmail receive mail at an address, as the last relay
in the chain. Without this, messages addressed to failmail will get
stored and ultimately summarized and sent back to itself. Now, some
metadata is stored with the message that changes the effective recipient
for summarization purposes.